==> ch/1. external_data_clustering.ipynb <==
# 전처리

import pandas as pd
import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings

warnings.filterwarnings('ignore')
tqdm.pandas()

# hyperparameter settings
tolerance = 0.8 # micro-aggregation level에서 몇%의 잘못된 결과를 허용하는지 여부 (0.5 ~ 1.0)
categorize = 3 # 연속형 변수를 몇 개의 양자화를 허용할 것인지 여부 (2 ~ 3)

df1 = pd.read_csv("internal.csv", encoding = "utf-8-sig")
df2 = pd.read_csv("integrated.csv", encoding = "utf-8-sig")
df3 = pd.read_csv("mediator.csv", encoding = "utf-8-sig")

for i in tqdm(df2.columns):
    for w in df2[i]:
        try:
            if "기타" in w:
                print(w, " ", i)
                break
            elif "X" in w:
                print(w, " ", i)
                break
        except:
            pass

# 제외 컬럼
del df2["PET_ACTIVE_SENIOR"]
del df2["OTHER_CNT"]
del df2["OTHER_AMT"]
del df2["OTHER_RATE"]
del df2["PER_OCH_WEDD"]

# 기타 및 XX 제외
df2["JOB"] = [np.nan if i == "04_기타" else i for i in df2["JOB"]]
df2["INCOME"] = [np.nan if i == "04_XX" else i for i in df2["INCOME"]]
df2["TIME_SEG"] = [np.nan if i == "06.기타" else i for i in df2["TIME_SEG"]]

# 서브 정보 활용
del df2["MGPO_MAIN"]
del df2["MGPO_HOM"]
del df2["SGG_HOM"]
del df2["MGPO_OFFI"]
del df2["SGG_OFFI_SEG"]
del df2["SGG_MAIN"]
del df2["DONG_MAIN"]
del df2["MAIN_AREA"]

# AMT CNT
del df2["OTHER_CARE_CNT"]
del df2["CMPPT_CNT"]
del df2["ONLINE_CSMTIC_CNT"]
del df2["DRGSTR_CNT"]
del df2["ONLINE_CNT"]
del df2["NUTRI_CNT"]

# >= 3만 해당 스테이지로 가정
df2["PER_WEDD_EXP"] = [1 if i >= 1 else 0 for i in tqdm(df2["PER_WEDD_EXP"])]
df2["PER_CHDB_EXP"] = [1 if i >= 1 else 0 for i in tqdm(df2["PER_CHDB_EXP"])]
# df2["PER_OCH_WEDD"] = [1 if i >= 1 else 0 for i in tqdm(df2["PER_OCH_WEDD"])]
df2["PER_HOUSEWF"] = [1 if i >= 1 else 0 for i in tqdm(df2["PER_HOUSEWF"])]
df2["PER_ONEPER"] = [1 if i >= 1 else 0 for i in tqdm(df2["PER_ONEPER"])]
df2["PER_PET"] = [1 if i >= 1 else 0 for i in tqdm(df2["PER_PET"])]

seg = dict(zip(df3["INCS_NO"], df3["SEG_KEY"]))
df1["seg_key"] = df1["INCS_NO"].progress_apply(lambda x : seg[x])

seg = dict(zip(df1["seg_key"], df1["BEHAVIOR_LABEL"]))
df2["BEHAVIOR_LABEL"] = df2["SEG_KEY"].progress_apply(lambda x : seg[x])

temp = [df2.loc[i].isnull().sum() for i in tqdm(df2.index)]
df2["matched"] = [1 if i < 20 else 0 for i in tqdm(temp)]
df2 = df2[df2["matched"] == 1].reset_index(drop = True)
del df2["matched"]

binary_columns = ["PER_WEDD_EXP", "PER_CHDB_EXP", "PER_HOUSEWF", "PER_ONEPER", "PER_PET", "CSMTC_PRFR_BIT"]
categorical_columns = ["JOB", "INCOME", "LFSTG", "TIME_CSMTC_SEG", "TIME_CAFE_SEG", "TIME_SEG"]

numeric_columns = list(df2.columns[1:-1])
for i in binary_columns:
    numeric_columns.remove(i)
for i in categorical_columns:
    numeric_columns.remove(i)

for i in numeric_columns:
    df2[i] = [np.nan if float(w) == 0 else w for w in df2[i]]

seg = list(seg.keys())
info_result = pd.DataFrame(index = seg, columns = binary_columns)
info_result

# binary filter
bc = df2[binary_columns + ["SEG_KEY"]]
for i in tqdm(seg):
    temp = bc[bc["SEG_KEY"] == i].dropna().reset_index(drop = True)

    for w in binary_columns:
        if (np.mean(temp[w]) >= tolerance) & (len(temp.index) > 1):
            info_result[w].loc[i] = 1
        else:
            info_result[w].loc[i] = 0

# categorical filter
info_result[categorical_columns] = np.nan
cc = df2[categorical_columns + ["SEG_KEY"]]

for i in tqdm(seg):
    temp = cc[cc["SEG_KEY"] == i].dropna().reset_index(drop = True)

    for w in categorical_columns:
        try:
            if (temp[w].value_counts().iloc[0]/temp[w].value_counts().sum() >= tolerance) & (len(temp.index) > 1):
                info_result[w].loc[i] = temp[w].value_counts().index[0]
            else:
                info_result[w].loc[i] = 0
        except:
            info_result[w].loc[i] = 0

info_result

# # categorical filter
# info_result[categorical_columns] = np.nan
# cc = df2[categorical_columns + ["SEG_KEY"]]
# for i in tqdm(seg):
#     temp = cc[cc["SEG_KEY"] == i]

#     for w in categorical_columns:
#         if len(temp[w].unique()) == 1:
#             info_result[w].loc[i] = temp[w].unique()[0]
#         else:
#             info_result[w].loc[i] = 0

# info_result

info_result[numeric_columns] = np.nan
nc = df2[numeric_columns + ["SEG_KEY"]]

baseline = pd.DataFrame(index = numeric_columns, columns = ["mean", "std"])
for i in tqdm(numeric_columns):
    temp = df2[i].dropna().reset_index(drop = True)
    if type(temp[0]) != str:
        baseline.loc[i] = [np.mean(temp), np.std(temp)]
    else:
        baseline.loc[i] = [np.nan, np.nan]
baseline

if categorize == 2:
    for i in tqdm(seg):
        temp = nc[nc["SEG_KEY"] == i]
    
        for w in numeric_columns:
            if np.std(temp[w]) < baseline["std"].loc[w]:
                if np.mean(temp[w]) > baseline["mean"].loc[w]:
                    info_result[w].loc[i] = "평균 이상"
                else:
                    # info_result[w].loc[i] = "평균 이하"
                    info_result[w].loc[i] = 0
            else:
                info_result[w].loc[i] = 0

else:
    for i in tqdm(seg):
        temp = nc[nc["SEG_KEY"] == i]
    
        for w in numeric_columns:
            if np.std(temp[w]) < baseline["std"].loc[w]:
                if np.mean(temp[w]) > baseline["mean"].loc[w] + baseline["std"].loc[w]:
                    info_result[w].loc[i] = "상"
                elif np.mean(temp[w]) < baseline["mean"].loc[w] - baseline["std"].loc[w]:
                    # info_result[w].loc[i] = "하"
                    info_result[w].loc[i] = 0
                else:
                    # info_result[w].loc[i] = "중"
                    info_result[w].loc[i] = 0
            else:
                info_result[w].loc[i] = 0

# info_result.to_csv("information_matrix8.csv", encoding = "utf-8-sig")

info_result = pd.read_csv("information_matrix8.csv", encoding = "utf-8-sig", index_col = "Unnamed: 0")

def convert(val):
    if val == "0.0":
        return int(float(val))
    return val
info_result = info_result.applymap(convert)
info_result

seg = dict(zip(df1["seg_key"], df1["BEHAVIOR_LABEL"]))

info_result["seg"] = np.nan
for i in tqdm(info_result.index):
    info_result["seg"].loc[i] = seg[i]
info_result

from itertools import combinations
from collections import Counter

col = info_result.columns[:-1]
result = pd.DataFrame(index = info_result.index)
result["cluster"] = np.nan

for i in tqdm(range(33)):
    temp = info_result[info_result["seg"] == i]
    lists = []
    for w in temp.index:
        lit = []
        for p in col:
            if (temp[p].loc[w] == 0) or (temp[p].loc[w] == "0"):
                pass
            else:
                lit.append(p + " " + str(temp[p].loc[w]))
        lists.append(lit)
    
    lists = dict(zip(temp.index, lists))
    
    while True:
        comb_counter = Counter()
        lits = list(lists.values())
        for s in lits:
            comb_counter.update(combinations(s, 5))
        
        most_common_combinations = comb_counter.most_common(1)
    
        if len(most_common_combinations) == 0:
            break
        
        zero = most_common_combinations[0][0][0]
        one = most_common_combinations[0][0][1]
        two = most_common_combinations[0][0][2]
        three = most_common_combinations[0][0][3]
        four = most_common_combinations[0][0][4]
        
        erase = []
        for w in lists:
            if (zero in lists[w]) and (one in lists[w]) and (two in lists[w]) and (three in lists[w]) and (four in lists[w]):
                result["cluster"].loc[w] = most_common_combinations[0][0]
                erase.append(w)
            else:
                pass
                
        for w in erase:
            del lists[w]

result.to_csv("result8.csv", encoding = "utf-8-sig")
