==> ch/3. Integrating Internal and External.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings
from collections import Counter
import ast

warnings.filterwarnings('ignore')
tqdm.pandas()

# internal data setting
internal = pd.DataFrame(columns = ["customer_id", "cluster", "sequence", "names", "tfidf", "tfidf_name"])
for num in tqdm(range(33)):
    temp = pd.read_csv(f"C:/Users/amore/Changhyun/vector/result{num}.csv")
    tfidf = pd.read_csv(f"C:/Users/amore/Changhyun/named_tfidf_results{num}.csv")
    
    indice = []
    vc = temp["cluster"].value_counts()
    for i in vc.index:
        if vc.loc[i] >= 1000:
            indice.append(i)
    for i in indice:
        temp2 = temp[temp["cluster"] == i][["customer_id", "cluster", "sequence", "names"]]
        temp2["cluster"] = [str(num) + "-" + str(w) for w in temp2["cluster"]]
        temp2["tfidf"] = [list(tfidf["item" + str(i)]) for w in temp2["cluster"]]
        temp2["tfidf_name"] = [list(tfidf["name" + str(i)]) for w in temp2["cluster"]]
        internal = pd.concat([internal, temp2], axis = 0)

# external data setting
external = pd.read_csv("result8.csv")
temp = dict(external["cluster"].value_counts())
vc = {}
n = 0
for i in temp:
    if temp[i] >= 50:
        vc[i] = n
        n += 1
    else:
        pass
num = []
for i in external["cluster"]:
    try:
        num.append(vc[i])
    except:
        num.append(np.nan)
external["cluster_num"] = num
external = external.dropna().reset_index(drop = True)
external["cluster"] = external["cluster"].apply(lambda x : ast.literal_eval(x))
external["cluster_num"] = external["cluster_num"].apply(lambda x : int(x))

# 추가 external 처리
from itertools import combinations
from collections import Counter

def find_most_common_subset(sets):
    # 원소의 개수가 4개 이상인 set만 필터링합니다.
    filtered_sets = [s for s in sets if len(s) >= 4]
    
    # 3개 부분집합들을 저장하기 위한 Counter를 정의합니다.
    subset_counter = Counter()
    
    # 모든 부분집합을 생성하여 카운팅합니다.
    for s in filtered_sets:
        for subset in combinations(s, 3):
            subset_counter[subset] += 1
    
    # 가장 많이 카운팅 된 3개 부분집합을 찾습니다.
    if subset_counter:
        most_common_subset, count = subset_counter.most_common(1)[0]
        return set(most_common_subset)
    else:
        return None

external2 = pd.DataFrame(columns = ["Unnamed: 0", "cluster", "cluster_num"])
idx = 0
n = 0
p = 1
while len(external.index) != 0:
    temp = find_most_common_subset(external["cluster"])
    left = []
    for a, i in enumerate(external["cluster"]):
        if temp.issubset(set(i)):
            left.append(0)
            external2.loc[idx] = [external["Unnamed: 0"].loc[a], tuple(temp), n]
            idx += 1
            if idx%1000 == 0:
                print(f"{p * 1000}번째 인덱스 달성!")
                p += 1
            else:
                pass
        else:
            left.append(1)
    external["left"] = left
    external = external[external["left"] == 1].reset_index(drop = True)
    del external["left"]
    n += 1
print(external2["cluster_num"].value_counts())
external = external2.copy()

# merge
med = pd.read_csv("mediator.csv")
med = dict(zip(med["INCS_NO"], med["SEG_KEY"]))

seg_key = []
external_cluster = []
external_cluster_num = []
for i in tqdm(internal["customer_id"]):
    seg_key.append(med[i])
    try:
        external_cluster.append(external[external["Unnamed: 0"] == med[i]]["cluster"].iloc[0])
    except:
        external_cluster.append(np.nan)
    
    try:
        external_cluster_num.append(external[external["Unnamed: 0"] == med[i]]["cluster_num"].iloc[0])
    except:
        external_cluster_num.append(np.nan)

internal["seg_key"] = seg_key
internal["external_cluster"] = external_cluster
internal["external_cluster_num"] = external_cluster_num

internal = internal.dropna().reset_index(drop = True)

thres = internal.groupby(["cluster", "external_cluster_num"])["customer_id"].count().sort_values(ascending = False)
idx = []
for i in thres.index:
    if thres.loc[i] >= 32:
        idx.append(i)

final = pd.DataFrame(columns = internal.columns)
for a, b in tqdm(idx):
    temp = internal[(internal["cluster"] == a) & (internal["external_cluster_num"] == b)]
    final = pd.concat([final, temp], axis = 0)
final = final.reset_index(drop = True)
final.to_csv("merged_data.csv") # 내부 x 외부

# 내부 추출
final = pd.read_csv("merged_data.csv", index_col = 0)
in_ex = list(final["customer_id"])

internal["filter"] = [1 if i in in_ex else 0 for i in tqdm(internal["customer_id"])]
internal = internal[internal["filter"] == 0].reset_index(drop = True)

in_ex = list(final["cluster"].unique())
internal["filter"] = [1 if i in in_ex else 0 for i in tqdm(internal["cluster"])]
internal = internal[internal["filter"] == 1].reset_index(drop = True)

vc = final["cluster"].value_counts()
internal_only = pd.DataFrame(columns = internal.columns)
for i in tqdm(in_ex):
    temp = internal[internal["cluster"] == i].sample(frac = 1).reset_index(drop = True).loc[:vc[i]]
    internal_only = pd.concat([internal_only, temp], axis = 0)
internal_only = internal_only.reset_index(drop = True)
internal_only

internal_only.to_csv("internal_only.csv", encoding = "utf-8-sig")
