==> ch/4. Defining characteristics.ipynb <==
import pandas as pd

teach_columns = {
    "TIME_CSMTC_SEG E.18-22시" : "저녁 시간에 화장품을 구매하는 고객",
    'INCOME 01_중하계층' : "저소득 고객",
    'INCOME 02_중상계층' : "고소득 고객",
    'ONLINE_CSMTIC_RATE 상' : "온라인으로 화장품을 많이 구매한 고객",
    'DRGSTR_RATE 상' : "drugstore 화장품을 많이 구매한 고객",
    'TIME_SEG 02.오전' : "주요 소비가 오전 시간대인 고객",
    "TIME_SEG 05.야간" : "주요 소비가 야간 시간대인 고객",
    'TIME_CAFE_SEG B.06-10시' : "아침 시간에 카페를 이용 빈도가 높은 고객",
    'TIME_CAFE_SEG B.10-14시' : "점심 시간에 카페를 이용 빈도가 높은 고객",
    'TIME_CAFE_SEG C.14-18시' : "오후 시간에 카페 이용 빈도가 높은 고객",
    'PER_ERADP 상' : "얼리어답터 성향이 있는 고객",
    'LFSTG 01.미혼추정' : "미혼으로 추정되는 고객",
    'PER_DOM_TRV 상' : "국내 여행에 관심이 많은 고객",
    'PER_FRN_TRV 상' : "해외 여행에 관심이 많은 고객",
    'PER_ONLINE 상' : "온라인 구매를 많이 하는 고객",
    'ONLINE_RATE 상' : "온라인 구매가 많은 고객",
    'SKINCARE_CNT_6M 상' : "스킨케어 제품을 선호하는 고객",
    'SKINCARE_CNT_1Y 상' : "스킨케어 제품을 선호하는 고객",
    'PER_PRC_SNSV 상' : "가격에 민감한 고객",
    'PER_INDR_SPRTS 상' : "실내 운동을 많이 하는 고객",
    'PER_OUTDOOR 상' : "외부 활동을 많이 하는 고객",
    'PER_DRIVE 상' : "운전을 많이 하는 고객",
    'PER_REMM 상' : "명품을 많이 구매하는 고객",
    'PER_ONGAME 상' : "온라인 게임을 많이 하는 고객",
    'PER_MDLV 상' : "배달 식품을 자주 먹는 고객",
    'PER_BENEFIT 상' : "혜택에 반응하는 고객",
    'PER_OTT 상' : "OTT 서비스에 관심이 많은 고객",
    'JOB 02_회사원' : "회사원 고객",
    "JOB 01_전문직" : "전문직 고객",
    'PER_PRMM_FASHION 상' : "명품 의류를 많이 구매하는 고객",
    'LFSTG 05.성인자녀추정' : "성인 자녀를 둔 고객",
    'PER_CAFE 상' : "카페를 많이 이용하는 고객",
    "PER_HLTHY_FOOD 상" : "건강식을 많이 구매하는 고객",
    "PER_ORGNIC_FOOD 상" : "유기농 제품을 많이 구매하는 고객",
    "NUTRI_CNT" : "건강 보조 식품을 많이 구매하는 고객",
    "NUTRI_AMT" : "건강 보조 식품을 많이 구매하는 고객",
}

desc = pd.read_csv("description.csv")
desc = desc.dropna(subset = "BASICDESC").reset_index(drop = True)
desc = desc.drop_duplicates(subset = "PRDNM").reset_index(drop = True)
desc = dict(zip(desc["SAPPRDCD"], desc["BASICDESC"]))
# desc

import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings
from collections import Counter
import ast

warnings.filterwarnings('ignore')
tqdm.pandas()

import requests
import base64
import json

# Azure GPT API 설정
GPT4V_ENDPOINT = "https://apeus-cst-openai.openai.azure.com/openai/deployments/gpt4o/chat/completions?api-version=2024-02-15-preview"
GPT4V_KEY = "fb73c5cdb51d4167b9eb41345d059bfe"

# 헤더 설정
headers = {
    "Content-Type": "application/json;",
    "api-key": GPT4V_KEY,
}

def respond(payload):
    response = requests.post(GPT4V_ENDPOINT, headers=headers, data=json.dumps(payload))
    
    if response.status_code == 200:
        result = response.json()
        return result['choices'][0]['message']['content']
        
    else:
        return f"Error {response.status_code}: {response.text}"

df = pd.read_csv("merged_data.csv", encoding = "utf-8-sig", index_col = "Unnamed: 0")
df["sequence"] = df["sequence"].apply(lambda x : ast.literal_eval(x))
df["names"] = df["names"].apply(lambda x : ast.literal_eval(x))
df["tfidf"] = df["tfidf"].apply(lambda x : ast.literal_eval(x))
df["tfidf_name"] = df["tfidf_name"].apply(lambda x : ast.literal_eval(x))
df["external_cluster"] = df["external_cluster"].apply(lambda x : ast.literal_eval(x))

# cf = pd.read_csv("internal_only.csv", encoding = "utf-8-sig", index_col = "Unnamed: 0")
# cf["sequence"] = cf["sequence"].apply(lambda x : ast.literal_eval(x))
# cf["names"] = cf["names"].apply(lambda x : ast.literal_eval(x))
# cf["tfidf"] = cf["tfidf"].apply(lambda x : ast.literal_eval(x))
# cf["tfidf_name"] = cf["tfidf_name"].apply(lambda x : ast.literal_eval(x))

review = pd.read_csv("item2.csv")
review = review.drop_duplicates(subset = "RPRS_PRD_CD").reset_index(drop = True)
review = dict(zip(review["RPRS_PRD_CD"], review["RPRS_PRD_NM"]))

def id_to_name(key):
    try:
        return review[int(float(key))]
    except:
        return key

names = []
for i in df["sequence"]:
    names.append([id_to_name(w) for w in i])
df["names"] = names

tfidf = []
for i in df["tfidf"]:
    tfidf.append([id_to_name(w) for w in i])
df["tfidf_name"] = tfidf
df

# clu = df.groupby(["cluster", "external_cluster_num"])[["tfidf_name", "external_cluster"]].first()
# clu.index.names = ["internal_cluster_num", "external_cluster_num"]
# clu["cluster"] = range(len(clu["tfidf_name"]))

# change = []
# for i in clu["tfidf_name"]:
#     temp = []
#     for w in i:
#         temp.append(w.replace("IF.", "").replace("이니스프리", "").strip())
#     change.append(temp)
# clu["tfidf_name"] = change

clu = df.groupby(["cluster", "external_cluster_num"])[["tfidf", "external_cluster"]].first()
clu.index.names = ["internal_cluster_num", "external_cluster_num"]
clu["cluster"] = range(len(clu["tfidf"]))

# clu2 = cf.groupby("cluster")["tfidf"].first()
# clu2.index.names = ["internal_cluster_num"]

# description 새로 받기
retrieve = {}
description = {}
for num in tqdm(df.index):
    temp1 = df["tfidf"].loc[num]
    temp2 = df["tfidf_name"].loc[num]
    for a, i in enumerate(temp1):
        retrieve[i] = str(temp2[a]).replace("IF.", "").replace("이니스프리", "").strip()
        try:
            description[retrieve[i]] = desc[i]
        except:
            pass

# 소구점 찾기 (internal only)
result = []
internal_info = []

idx = clu.index.get_level_values("internal_cluster_num").unique()

for temp in tqdm(idx):

    system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
    user1 = """
    내가 제공하는 고객 정보를 활용해서, 이 고객이 어떤 특징을 가지고 있는지 한 문장으로 정의하고, 어떤 정보를 보고 그렇게 정의했는지 그 번호를 알려줘.
    정보는 1개 - 3개 활용해 줘. 각 제품이 어떤 제품인지 알기 위해서 '제품 특징'을 참고해.
    단순히 사용한 정보들을 나열하지 말고, 정보들을 유기적으로 결합해서 고객 특징을 만들어야 해.
    
    우리가 원하는 답변의 예시는 다음과 같아: '바쁜 일상 속에서도 건강한 일상을 살기 위해 이른 아침부터 자기관리와 건강한 루틴을 실천하는 고객. 1번, 3번 정보.'
    """
    assistant1 = "알겠습니다. 정보와 제품 특징을 제공해 주세요."
    
    internal = ""
    tp = list(clu["tfidf"].loc[temp])[0]
    internal_info.append(tp)
    for a, i in enumerate(tp):
        try:
            internal += f"{a}번 정보: 이 고객은 과거에 {retrieve[i]}를 구매한 적이 있어.\n제품 특징: {description[retrieve[i]]}\n"
        except:
            internal += f"{a}번 정보: 이 고객은 과거에 {retrieve[i]}를 구매한 적이 있어.\n"

    payload = {
        "messages": [
            {"role": "system", "content": system1},
            {"role": "user", "content": user1},
            {"role": "assistant", "content": assistant1},
            {"role": "user", "content": internal}]}

    decision = True
    while decision:
        try:
            result.append(respond(payload))
            decision = False
        except:
            pass

regexp = re.compile("[0-9].+번 정보")
items = []
for a, i in enumerate(result):
    temp = regexp.findall(i)[0].split("번")
    temp2 = internal_info[a]
    item = []
    for w in temp[:-1]:
        item.append(temp2[int(w[-1])])
    items.append(item)

internal_df = pd.DataFrame({"internal_cluster" : list(idx), "focus_item" : items, "characteristics" : result})
internal_df

# # 소구점 찾기 (internal + external)
# result = []
# internal_info = []
# external_info = []

# for idx, temp in tqdm(enumerate(clu["tfidf"])):

#     system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
#     user1 = """
#     내가 제공하는 고객 정보를 활용해서, 이 고객이 어떤 특징을 가지고 있는지 한 문장으로 정의하고, 어떤 정보를 보고 그렇게 정의했는지 그 번호를 알려줘.
#     내부 정보는 1개 - 3개 활용할 수 있어. 각 제품이 어떤 제품인지 알기 위해서 '제품 특징'을 참고해.
#     외부 정보는 1개 - 3개 활용할 수 있어. 총 활용할 수 있는 정보에는 제한이 없어.
#     단순히 사용한 정보들을 나열하지 말고, 정보들을 유기적으로 결합해서 고객 특징을 만들어야 해.
#     우리가 원하는 답변의 예시는 다음과 같아:
#     '바쁜 일상 속에서도 건강한 일상을 살기 위해 이른 아침부터 자기관리와 건강한 루틴을 실천하는 고객. 내부 1번, 내부 3번, 외부 2번 정보.'
#     '주로 온라인에 몰입되는 도파민 넘치는 생활습관을 가지고 있는 고객. 내부 3번, 외부 0번, 외부 2번 정보.
#     '이른 아침부터 하루를 시작하지만, 늘 시간과 에너지가 부족해 외식이나 차려먹는 밥 대신 배달음식을 시켜먹는 고객. 내부 0번 정보'
#     """
#     assistant1 = "알겠습니다. 정보와 제품 특징을 제공해 주세요."
#     user2 = """
#     외부 0번 정보: 이 고객은 회사원이야.
#     외부 1번 정보: 이 고객은 카페에 지출하는 비용이 많아.
#     외부 2번 정보: 이 고객은 6시 - 12시 사이에 카페를 주로 사용해.
#     외부 3번 정보: 이 고객은 건강식에 대한 지출이 많아.
#     내부 0번 정보: 이 고객은 과거에 아이페놀썬크림을 구매한 적이 있어.
#     내부 1번 정보: 이 고객은 과거에 그린티영양크림을 구매한 적이 있어.
#     """
#     assistant2 = "이른 아침부터 바쁘고 스트레스가 많은 일상을 살며 카페인이 있어야만 사는 고객. 내부 1번, 외부 0번, 외부 2번, 외부 3번 정보."
    
#     internal = ""
#     internal_info.append(temp)
#     for a, i in enumerate(temp):
#         try:
#             internal += f"내부 {a}번 정보: 이 고객은 과거에 {retrieve[i]}를 구매한 적이 있어.\n제품 특징: {description[retrieve[i]]}\n"
#         except:
#             internal += f"내부 {a}번 정보: 이 고객은 과거에 {retrieve[i]}를 구매한 적이 있어.\n"

#     external = ""
#     tp = clu["external_cluster"].iloc[idx]
#     external_info.append(tp)
#     for a, i in enumerate(tp):
#         external += f"외부 {a}번 정보: 이 고객은 {teach_columns[i]}이야.\n"
    
#     payload = {
#         "messages": [
#             {"role": "system", "content": system1},
#             {"role": "user", "content": user1},
#             {"role": "assistant", "content": assistant1},
#             {"role": "user", "content": user2},
#             {"role": "assistant", "content": assistant2},
#             {"role": "user", "content": external + internal}]}

#     decision = True
#     while decision:
#         try:
#             result.append(respond(payload))
#             decision = False
#         except:
#             pass

# internal_regexp = re.compile("내부 [0-9]+번")
# external_regexp = re.compile("외부 [0-9]+번")

# items1 = []
# items2 = []

# for a, i in enumerate(result):
#     temp1 = internal_regexp.findall(i)
#     temp1 = [w[3:-1] for w in temp1]
    
#     temp2 = external_regexp.findall(i)
#     temp2 = [w[3:-1] for w in temp2]

#     check1 = internal_info[a]
#     check2 = external_info[a]

#     item1 = []
#     for w in temp1:
#         item1.append(check1[int(w)])

#     item2 = []
#     for w in temp2:
#         item2.append(check2[int(w)])

#     items1.append(item1)
#     items2.append(item2)
    
# external_df = pd.DataFrame({"internal_item" : items1, "external_item" : items2, "characteristics" : result}, index = clu.index)
# external_df

# 소구점 찾기 (external only)
result = []
external_info = []

for idx, temp in tqdm(enumerate(clu["tfidf"])):

    system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
    user1 = """
    내가 제공하는 고객 정보를 활용해서, 이 고객이 어떤 특징을 가지고 있는지 한 문장으로 정의하고, 어떤 정보를 보고 그렇게 정의했는지 그 번호를 알려줘.
    정보는 1개 - 3개 활용할 수 있어. 단순히 사용한 정보들을 나열하지 말고, 정보들을 유기적으로 결합해서 고객 특징을 만들어야 해.
    우리가 원하는 답변의 예시는 다음과 같아:
    '바쁜 일상 속에서도 건강한 일상을 살기 위해 이른 아침부터 자기관리와 건강한 루틴을 실천하는 고객. 2번 정보.'
    '주로 온라인에 몰입되는 도파민 넘치는 생활습관을 가지고 있는 고객. 0번, 2번 정보.
    '이른 아침부터 하루를 시작하지만, 늘 시간과 에너지가 부족해 외식이나 차려먹는 밥 대신 배달음식을 시켜먹는 고객. 0번, 1번, 3번 정보'
    """
    assistant1 = "알겠습니다. 정보와 제품 특징을 제공해 주세요."
    
    user2 = """
    0번 정보: 이 고객은 회사원이야.
    1번 정보: 이 고객은 카페에 지출하는 비용이 많아.
    2번 정보: 이 고객은 6시 - 12시 사이에 카페를 주로 사용해.
    3번 정보: 이 고객은 건강식에 대한 지출이 많아.
    """
    assistant2 = "이른 아침부터 바쁘고 스트레스가 많은 일상을 살며 카페인이 있어야만 사는 고객. 0번, 2번, 3번 정보."

    user3 = """
    0번 정보: 이 고객은 OTT 사용 시간이 많아.
    1번 정보: 이 고객은 온라인 쇼핑에 지출하는 비용이 많아.
    2번 정보: 이 고객은 최신 트렌드 패션에 대한 관심이 높아.
    3번 정보: 이 고객은 외부 활동을 많이 해.
    """
    assistant3 = "늘 새로운 트렌드를 쫒고, 전자기기기가 새로 나오면 먼저 써봐야 하고, OTT로 문화생활 섭렵하기 바쁘고 누구보다 발빠르게 소비하는 고객. 0번, 1번, 2번 정보."

    external = ""
    tp = clu["external_cluster"].iloc[idx]
    external_info.append(tp)
    for a, i in enumerate(tp):
        external += f"{a}번 정보: 이 고객은 {teach_columns[i]}이야.\n"
    
    payload = {
        "messages": [
            {"role": "system", "content": system1},
            {"role": "user", "content": user1},
            {"role": "assistant", "content": assistant1},
            {"role": "user", "content": user2},
            {"role": "assistant", "content": assistant2},
            {"role": "user", "content": external}]}

    decision = True
    while decision:
        try:
            result.append(respond(payload))
            decision = False
        except:
            pass

items = []
for a, i in enumerate(result):
    temp = regexp.findall(i)[0].split("번")
    temp2 = external_info[a]
    item = []
    for w in temp[:-1]:
        item.append(temp2[int(w[-1])])
    items.append(item)
    
external_df = pd.DataFrame({"external_item" : items, "characteristics" : result}, index = clu.index)
external_df

internal_df.to_csv("internal_characteristics.csv", encoding = "utf-8-sig")
external_df.to_csv("external_characteristics.csv", encoding = "utf-8-sig")
