==> ch/5. Targeting.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings
from collections import Counter
import ast
import pickle

warnings.filterwarnings('ignore')
tqdm.pandas()

import requests
import base64
import json

# Azure GPT API 설정
GPT4V_ENDPOINT = "https://apeus-cst-openai.openai.azure.com/openai/deployments/gpt4o/chat/completions?api-version=2024-02-15-preview"
GPT4V_KEY = "fb73c5cdb51d4167b9eb41345d059bfe"

# 헤더 설정
headers = {
    "Content-Type": "application/json;",
    "api-key": GPT4V_KEY,
}

def respond(payload):
    response = requests.post(GPT4V_ENDPOINT, headers=headers, data=json.dumps(payload))
    
    if response.status_code == 200:
        result = response.json()
        return result['choices'][0]['message']['content']
        
    else:
        return f"Error {response.status_code}: {response.text}"

internal_df = pd.read_csv("internal_characteristics.csv", encoding = "utf-8-sig", index_col = 0)
internal_df["focus_item"] = internal_df["focus_item"].apply(lambda x : ast.literal_eval(x))

external_df = pd.read_csv("external_characteristics.csv", encoding = "utf-8-sig")
# external_df["internal_item"] = external_df["internal_item"].apply(lambda x : ast.literal_eval(x))
external_df["external_item"] = external_df["external_item"].apply(lambda x : ast.literal_eval(x))
idf = dict(zip(internal_df["internal_cluster"], internal_df["characteristics"]))
external_df["internal_characteristics"] = [idf[x] for x in external_df["internal_cluster_num"]]

# 내부 데이터로 피부 고민 찾기
internal_gomin = []

system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
    
user1 = """
나는 너에게 고객의 특징을 알려줄거야. 하지만 지금 정의된 특징들이 너무 고객에게 이야기하기에는 너무 직접적인 내용을 담고 있어.
그 특징을 가진 고객이 어떤 피부 고민을 가지고 있을지, 또는 어떤 화장품 제품을 원할지, 아래 조건들을 참고해서 말해줘.
조건 1. 고객의 특징이 언급하고 있는 정보들을 반드시 모두 사용해야만 해.
조건 2. 환경 보호, 자연 보호, 클린 뷰티는 절대로 포함되어서는 안 돼.
조건 3. 주어진 정보들을 직접적으로 사용하기보다, 정보들을 유기적으로 조합해서 감성적인 특징을 제시해줬으면 좋겠어.
"""
assistant1 = "알겠습니다. 고객의 특징을 알려주세요."

user2 = "한 군데에만 있으면 좀이 쑤셔 여기저기 돌아다녀야 직성이 풀리고, 여행 시 들고다닐 간편한 아이템을 원하는 고객."
assistant2 = "휴대성과 사용성이 간편한 제품을 원함. 잦은 환경변화로 인해 피부 트러블 올라 오는 경우가 많아 늘 사용하던 제품이나 저자극의 순한 제품을 원함."

user3 = "에이블리/지그재그에서 패션 쇼핑하고 올리브영을 자주가며 OTT를 구독하고 문화생활을 많이 하는 고객. 카드사, 온라인 할인 혜택을 잘 사용하는 고객."
assistant3 = "피지 분비가 많은 시기, 호르몬 변화에 의한 피부 트러블이 자주 올라옴. 여드름 흉터, 블랙헤드 등 트러블 자국과 고르지 않은 피부결에 대해 고민이 높음."

user4 = "뷰티관심러로서 뷰티 관리를 철저히 하는 고객. 피부과, 피부 에스테틱을 자주 다니며 헤어/네일 등 미용 관리 역시 높은 관심을 가지고 있는 고객. 심지어 이너뷰티까지 중요성까지 알고 있기에 샐러드 등 건강식단을 중요시하며, 실내운동도 꾸준히 하는 고객."
assistant4 = "일상 생활 중 자극을 받아 생겨난 트러블, 붉은 기, 전반적인 피부 장벽 및 컨디션 저하를 해결하고 싶어함. 메이크업이 잘 먹는 최적의 상태의 피부(촉촉하고 수분 가득하고, 피부결이 고르게 정돈된 피부)를 원함"

for i in tqdm(internal_df["characteristics"]):

    payload = {
        "messages": [
            {"role": "system", "content": system1},
            {"role": "user", "content": user1},
            {"role": "assistant", "content": assistant1},
            {"role": "user", "content": user2},
            {"role": "assistant", "content": assistant2},
            {"role": "user", "content": user3},
            {"role": "assistant", "content": assistant3},
            {"role": "user", "content": user4},
            {"role": "assistant", "content": assistant4},
            {"role": "user", "content": i}]}
    
    decision = True
    while decision:
        try:
            r = respond(payload)
            if "죄송" in r:
                pass
            else:
                internal_gomin.append(r)
                decision = False
        except:
            pass

internal_df["concern"] = internal_gomin

# 외부 데이터로 피부 고민 찾기
external_gomin = []

for a, i in tqdm(enumerate(external_df["characteristics"])):

    payload = {
        "messages": [
            {"role": "system", "content": system1},
            {"role": "user", "content": user1},
            {"role": "assistant", "content": assistant1},
            {"role": "user", "content": user2},
            {"role": "assistant", "content": assistant2},
            {"role": "user", "content": user3},
            {"role": "assistant", "content": assistant3},
            {"role": "user", "content": i + " " + external_df["internal_characteristics"].loc[a]}]}
    
    decision = True
    while decision:
        try:
            r = respond(payload)
            if "죄송" in r:
                pass
            else:
                external_gomin.append(r)
                decision = False
        except:
            pass

external_df["concern"] = external_gomin

# df = pd.read_csv("item2.csv")

# target_items = {"그린티씨드세럼" : 111705279, "그린티씨드크림" : 111770419,
#                 "레티놀시카앰플" : 131173569, "레티놀PDRN앰플" : 131175032,
#                 "비타C엔자임세럼" : 131173960}

# result = {}

# for i in tqdm(target_items):
#     temp = list(df[df['RPRS_PRD_CD'] == target_items[i]]["GOODCNT"])

#     system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
#     user1 = f"""
#     나는 너에게 {i}에 대한 고객 리뷰들을 제공해 줄거야. 그 리뷰들을 요약해서, {i}가 어떤 제품인지 자세하게 말해줘.
#     """
#     assistant1 = "알겠습니다. 리뷰를 제공해 주세요."

#     payload = {
#         "messages": [
#             {"role": "system", "content": system1},
#             {"role": "user", "content": user1},
#             {"role": "assistant", "content": assistant1},
#             {"role": "user", "content": str(temp)}]}

#     decision = True
#     while decision:
#         try:
#             result[i] = respond(payload)
#             decision = False
#         except:
#             pass
    
# with open(file='recommended_contents.pickle', mode='wb') as f:
#     pickle.dump(result, f)

cp = {
    "그린티씨드세럼" : "대한민국 판매 1위 수분에센스. 단숨에 속건조 해결",
    "그린티씨드크림" : "외부자극으로 성난 피부 수분 진정. 재구매율 1위. 써 본 사람은 다 아는 확실한 수분 진정 효과. 수분을 빈틈 없이 꽉 채워 화장이 잘 먹는 수분크림",
    "레티놀시카앰플" : "국민 트러블 앰플. 갓 생긴 트러블부터 모공, 잡티, 흔적까지 트러블 끝장 케어. 반복되는 트러블 고리의 악순환을 끊는 #레티놀시카. 트러블, 모공, 잡티, 흔적 고민 지우고 깨끗하고 매끈매끈한 깐달걀 피부결",
    "레티놀PDRN앰플" : "모공결부터 속광까지 #바르는 스킨 부스터. 매끈한 모공결 & 촘촘한 피부광의 스킨부스터 앰플. 주름, 모공, 탄력 복잡한 피부 고민 이거 하나로 끝! 당신의 피부를 전방위적으로 개선합니다."
}

# 내부 데이터 타겟팅

internal_result = []
for i in tqdm(internal_df["concern"]):
    
    system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
    user1 = f"""
    나는 너에게 특정 고객군의 피부 고민을 알려줄거야.
    그 특징을 바탕으로, [그린티씨드세럼, 그린티씨드크림, 레티놀시카앰플, 레티놀PDRN앰플] 중에서 가장 적절한 제품 하나를 추천해줘.
    그리고 그 추천이 얼마나 자신있는지 0% - 100%로 표시해준 후, 추천을 한 이유도 함께 설명해줘.
    우리가 선호하는 답변의 포맷은 다음과 같아: '레티놀시카앰플; 75%; 이 고객군은 피부가 예민하기 때문에, 순한 제품이 어울릴 것으로 예상됩니다.'
    각 제품에 대한 설명은 'Items'를 참조해.
    Items: {cp}
    """
    assistant1 = "알겠습니다. 고객군의 특징을 제공해 주세요."

    payload = {
        "messages": [
            {"role": "system", "content": system1},
            {"role": "user", "content": user1},
            {"role": "assistant", "content": assistant1},
            {"role": "user", "content": i}]}

    decision = True
    while decision:
        try:
            internal_result.append(respond(payload))
            decision = False
        except:
            pass

one = []
two = []
three = []
for i in internal_result:
    temp = i.split(";")
    one.append(temp[0])
    two.append(temp[1])
    three.append(temp[2])

internal_df["recommended_item"] = one
internal_df["confidence"] = two
internal_df["reason_of_recommendation"] = three
internal_df

# 내부 + 외부 데이터 타겟팅

external_result = []
for i in tqdm(external_df["concern"]):
    
    system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
    user1 = f"""
    나는 너에게 특정 고객군의 피부 고민을 알려줄거야.
    그 특징을 바탕으로, [그린티씨드세럼, 그린티씨드크림, 레티놀시카앰플, 레티놀PDRN앰플] 중에서 가장 적절한 제품 하나를 추천해줘.
    그리고 그 추천이 얼마나 자신있는지 0% - 100%로 표시해준 후, 추천을 한 이유도 함께 설명해줘.
    우리가 선호하는 답변의 포맷은 다음과 같아: '레티놀시카앰플; 75%; 이 고객군은 피부가 예민하기 때문에, 순한 제품이 어울릴 것으로 예상됩니다.'
    각 제품에 대한 설명은 'Items'를 참조해.
    Items: {cp}
    """
    assistant1 = "알겠습니다. 고객군의 특징을 제공해 주세요."

    payload = {
        "messages": [
            {"role": "system", "content": system1},
            {"role": "user", "content": user1},
            {"role": "assistant", "content": assistant1},
            {"role": "user", "content": i}]}

    decision = True
    while decision:
        try:
            external_result.append(respond(payload))
            decision = False
        except:
            pass

one = []
two = []
three = []
for i in external_result:
    temp = i.split(";")
    one.append(temp[0])
    two.append(temp[1])
    three.append(temp[2])

external_df["recommended_item"] = one
external_df["confidence"] = two
external_df["reason_of_recommendation"] = three
external_df

internal_df.to_csv("internal_characteristics.csv", encoding = "utf-8-sig")
external_df.to_csv("external_characteristics.csv", encoding = "utf-8-sig")

print(internal_df["recommended_item"].unique())
print(external_df["recommended_item"].unique())
