==> ch/6. Message generation.ipynb <==
import pickle

rec_contex = {
    "가성비를 중시하는 고객" : "빅세일 기간에 최대 60% 할인이 이루어진다는 사실을 강조할 수 있음",
    "그린티씨드크림을 추천하는 경우" : "대용량 구매 시 1+1 혜택이 있음",
    "여행이나 야외 활동을 많이 하는 고객" : "작은 용량의 제품을 추천하는 것이 권장됨",
    "그린티씨드세럼을 추천하는 경우" : "리필을 할 수 있음을 강조할 수 있음",
    "트렌드에 민감한 고객" : "구매 시 '쿠키 키링 파우치' 증정을 강조할 수 있음",
    "카페를 좋아하는 고객" : "연말에 화장품을 구매하는 것을 강조할 수 있음"
}

with open(file='recommended_contents.pickle', mode='rb') as f:
    rc = pickle.load(f)

cp = {
    "그린티씨드세럼" : "대한민국 판매 1위 수분에센스. 단숨에 속건조 해결",
    "그린티씨드크림" : "외부자극으로 성난 피부 수분 진정. 재구매율 1위. 써 본 사람은 다 아는 확실한 수분 진정 효과. 수분을 빈틈 없이 꽉 채워 화장이 잘 먹는 수분크림",
    "레티놀시카앰플" : "국민 트러블 앰플. 갓 생긴 트러블부터 모공, 잡티, 흔적까지 트러블 끝장 케어. 반복되는 트러블 고리의 악순환을 끊는 #레티놀시카. 트러블, 모공, 잡티, 흔적 고민 지우고 깨끗하고 매끈매끈한 깐달걀 피부결",
    "레티놀PDRN앰플" : "모공결부터 속광까지 #바르는 스킨 부스터. 매끈한 모공결 & 촘촘한 피부광의 스킨부스터 앰플. 주름, 모공, 탄력 복잡한 피부 고민 이거 하나로 끝! 당신의 피부를 전방위적으로 개선합니다."
}

import pandas as pd
import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings
from collections import Counter
import ast
from random import shuffle

warnings.filterwarnings('ignore')
tqdm.pandas()

import requests
import base64
import json

# Azure GPT API 설정
GPT4V_ENDPOINT = "https://apeus-cst-openai.openai.azure.com/openai/deployments/gpt4o/chat/completions?api-version=2024-02-15-preview"
GPT4V_KEY = "fb73c5cdb51d4167b9eb41345d059bfe"

# 헤더 설정
headers = {
    "Content-Type": "application/json;",
    "api-key": GPT4V_KEY,
}

def respond(payload):
    response = requests.post(GPT4V_ENDPOINT, headers=headers, data=json.dumps(payload))
    
    if response.status_code == 200:
        result = response.json()
        return result['choices'][0]['message']['content']
        
    else:
        return f"Error {response.status_code}: {response.text}"

internal_df = pd.read_csv("internal_characteristics.csv", encoding = "utf-8-sig", index_col = 0)
internal_df["focus_item"] = internal_df["focus_item"].apply(lambda x : ast.literal_eval(x))

external_df = pd.read_csv("external_characteristics.csv", encoding = "utf-8-sig", index_col = 0)
external_df["external_item"] = external_df["external_item"].apply(lambda x : ast.literal_eval(x))

find = re.compile('그린티씨드세럼|그린티씨드크림|레티놀시카앰플|레티놀PDRN앰플|비타C엔자임세럼')
internal_df["recommended_item"] = internal_df["recommended_item"].apply(lambda x : find.search(x.replace(" ", ""))[0])
external_df["recommended_item"] = external_df["recommended_item"].apply(lambda x : find.search(x.replace(" ", ""))[0])

already_suggested = []

system1 = "너는 화장품 기업 소속의 마케팅 매니저야."
user1 = """
나는 너에게 (1) 고객의 특징, (2) 추천하는 제품, (3) 추천하는 제품의 특징, (4) 추천의 이유 (5) 추천 맥락, (6) 피해야 할 문구 스타일들을 제공해 줄거야.
아래 포맷과 조건, 예시를 따라서, 고객에게 이 제품을 추천하는 문구를 2-3 문장으로 디자인해줘.
피해야할 문구 스타일들과 가급적이면 다른 스타일의 문구를 만들어줘. 없으면 그냥 만들어도 돼.
포맷: (1) 고객의 특징에 따른 맥락 언급하기. (2) 고객의 맥락에 의거해서 제품 추천하기
조건: 고객의 특징은 너무 세세하게 말해선 안 돼. 기본적으로 2문장을 생성하되, 사용해야 할 정보가 많아지면 3문장을 생성할 수 있어.
예시: [가성비 있는 피부결 관리템을 찾으시나요? 순한 성분으로 속건조를 해결하는 그린티씨드크림을 만나보세요!;
      바쁜 일상, 간단하지만 확실한 안티에이징 케어가 필요하다면? 다양한 피부 고민을 동시에 해결하는 레티놀PDRN앰플을 만나보세요!]
"""
assistant1 = "알겠습니다. 정보를 제공해 주세요."

user2 = f"""
(1) 고객의 특징: 바쁜 회사 생활 속에서도 건강한 식습관을 유지하는 고객.
(2) 추천하는 제품: 그린티씨드세럼
(3) 추천하는 제품의 특징: {cp["그린티씨드세럼"]}
(4) 추천의 이유: 이 고객군은 자연을 사랑하고 건강한 생활 습관을 유지하려고 노력하는 사람들이기 때문에, 천연 성분이 포함된 순하고 촉촉한 제품이 어울릴 것으로 예상됩니다. 그린티씨드세럼은 수분 공급과 순한 성분 덕분에 민감한 피부에도 적합하며, 바쁜 일상 속에서도 빠르게 흡수되어 사용하기 편리합니다.
(5) 추천 맥락: {rec_contex}
(6) 피해야 할 문구 스타일들: {already_suggested}
"""

assistant2 = "칼퇴근 후 배달 음식 대신 그린티 어때요? 자극받은 피부 진정시켜주는 그린티 씨드크림을 만나보세요"

user3 = f"""
(1) 고객의 특징: 외국과 국내 여행에 관심이 많고, 피부와 뷰티 제품에도 신경을 쓰며, 다양한 혜택과 프로모션에 민감하게 반응하는 고객.
(2) 추천하는 제품: 그린티씨드크림
(3) 추천하는 제품의 특징: {cp["그린티씨드크림"]}
(4) 추천의 이유:  이 고객군은 여행을 좋아하고 피부와 뷰티 제품에 관심이 많아 위생적이고 휴대가 간편한 제품이 필요합니다. 그린티씨드크림은 튜브형 포장으로 여행 시 휴대하기 편리하며 피부에 강력한 수분을 공급해줍니다.
(5) 추천 맥락: {rec_contex}
(6) 피해야 할 문구 스타일들: {already_suggested}
"""

assistant3 = "연말 나들이에는 꼭 데려가세요. 휴대하기 편한 강력한 수분 공급템, 그린티씨드크림을 만나보세요"

def generate_message_internal(internal_df):
    internal_result = []
    for i in tqdm(internal_df.index):
        try:
            already_suggested = list(internal_df[internal_df.columns[7:]].loc[i])
        except:
            already_suggested = []

        temp = internal_df["concern"].loc[0].strip().split(".")
        shuffle(temp)
        
        user4 = f"""
        (1) 고객의 특징: {". ".join(temp).strip()}
        (2) 추천하는 제품: {internal_df["recommended_item"].loc[i]}
        (3) 추천하는 제품의 특징: {rc[internal_df["recommended_item"].loc[i]]}
        (4) 추천의 이유: {internal_df["reason_of_recommendation"].loc[i]}
        (5) 추천 맥락: {rec_contex}
        (6) 피해야 할 문구 스타일들: {already_suggested}
        """
    
        payload = {
            "messages": [
                {"role": "system", "content": system1},
                {"role": "user", "content": user1},
                {"role": "assistant", "content": assistant1},
                {"role": "user", "content": user2},
                {"role": "assistant", "content": assistant2},
                {"role": "user", "content": user3},
                {"role": "assistant", "content": assistant3},
                {"role": "user", "content": user4}]}
    
        decision = True
        while decision:
            try:
                internal_result.append(respond(payload))
                decision = False
            except:
                pass
    return internal_result

def generate_message_external(external_df):
    external_result = []
    for i in tqdm(external_df.index):
        try:
            already_suggested = list(external_df[external_df.columns[9:]].loc[i])
        except:
            already_suggested = []

        temp = external_df["concern"].loc[0].strip().split(".")
        shuffle(temp)
            
        user4 = f"""
        (1) 고객의 특징: {". ".join(temp).strip()}
        (2) 추천하는 제품: {external_df["recommended_item"].loc[i]}
        (3) 추천하는 제품의 특징: {rc[external_df["recommended_item"].loc[i]]}
        (4) 추천의 이유: {external_df["reason_of_recommendation"].loc[i]}
        (5) 추천 맥락: {rec_contex}
        (6) 피해야 할 문구 스타일들: {already_suggested}
        """
    
        payload = {
            "messages": [
                {"role": "system", "content": system1},
                {"role": "user", "content": user1},
                {"role": "assistant", "content": assistant1},
                {"role": "user", "content": user2},
                {"role": "assistant", "content": assistant2},
                {"role": "user", "content": user3},
                {"role": "assistant", "content": assistant3},
                {"role": "user", "content": user4}]}
    
        decision = True
        while decision:
            try:
                external_result.append(respond(payload))
                decision = False
            except:
                pass
                
    return external_result

for i in range(2):
    internal_df[f"message{i}"] = generate_message_internal(internal_df)
    external_df[f"message{i}"] = generate_message_internal(external_df)

# internal_df.to_csv("internal_characteristics.csv", encoding = "utf-8-sig")
# external_df.to_csv("external_characteristics.csv", encoding = "utf-8-sig")

internal_df

external_df
