==> ch/7. Filing.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings
from collections import Counter
import ast
import pickle
import random

warnings.filterwarnings('ignore')
tqdm.pandas()

internal_df = pd.read_csv("internal_characteristics.csv", index_col = 0)
internal_df["focus_item"] = internal_df["focus_item"].apply(lambda x : ast.literal_eval(x))

external_df = pd.read_csv("external_characteristics.csv", index_col = 0)
external_df["external_item"] = external_df["external_item"].apply(lambda x : ast.literal_eval(x))

df = pd.read_csv("merged_data.csv", encoding = "utf-8-sig", index_col = 0)
cf = pd.read_csv("internal_only.csv", encoding = "utf-8-sig", index_col = 0)

df = df[["customer_id", "cluster", "tfidf_name", "external_cluster_num", "external_cluster"]]
df.columns = ["customer_id", "internal_cluster", "internal_cluster_info", "external_cluster", "external_cluster_info"]

cf = cf[["customer_id", "cluster", "tfidf_name"]]
cf.columns = ["customer_id", "internal_cluster", "internal_cluster_info"]
cf["external_cluster"] = np.nan
cf["external_cluster_info"] = np.nan

df = pd.concat([df, cf], axis = 0).reset_index(drop = True)
df

# df = pd.read_csv("merged_data.csv")
# df = df[["customer_id", "cluster", "tfidf_name", "external_cluster_num", "external_cluster"]]
# df.columns = ["customer_id", "internal_cluster", "internal_cluster_info", "external_cluster", "external_cluster_info"]

# df["external_cluster"] = df["external_cluster"].apply(lambda x : "-" + str(int(x)))

# df["total_cluster"] = df["internal_cluster"] + df["external_cluster"]
# col = list(df.columns)
# col.append("group")
# final_df = pd.DataFrame(columns = col)
# for i in tqdm(list(df["total_cluster"].unique())):
#     temp = df[df["total_cluster"] == i]
#     rand = [random.choice([0, 1, 2, 3]) for i in temp["customer_id"]]
#     while len(set(rand)) != 4:
#         rand = [random.choice([0, 1, 2, 3]) for i in temp["customer_id"]]
#     temp["group"] = rand
#     final_df = pd.concat([final_df, temp], axis = 0)
# final_df = final_df.reset_index(drop = True)
# final_df["external_cluster"] = final_df["external_cluster"].apply(lambda x : int(x[1:]))

pd.read_csv("total_samples2.csv")

final_df
