==> ch/8. Final arrangement.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
import re
import warnings
from collections import Counter
import ast
import pickle
import random

warnings.filterwarnings('ignore')
tqdm.pandas()

df = pd.read_csv("final_sample_20241211.csv", index_col = 0)
df = df.drop_duplicates(subset = "customer_id", keep = "first")
df["ec"] = df["external_cluster"].apply(lambda x : "-" + str(x)[:-2])
df["total_cluster"] = df["internal_cluster"] + df["ec"]

col = list(df.columns)
col.append("group")
result = pd.DataFrame(columns = col)

lit = list(df["total_cluster"].unique())

for i in tqdm(lit):
    temp = df[df["total_cluster"] == i]
    rand = []
    while len(set(rand)) != 4:
        rand = [random.choice([0, 1, 2, 3]) for i in temp["customer_id"]]
    temp["group"] = rand
    result = pd.concat([result, temp], axis = 0)

result = result.reset_index(drop = True)
del result["ec"]
result.to_csv("total_sample_matching_20241117.csv")

result
